+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - `nstd_string_from_str` now returns `NSTDOptionalString`.
 - `nstd_string_clone` now returns `NSTDOptionalString`.
 ### `nstd.thread`
+- `nstd_thread_spawn[_with_desc]` now takes `NSTDOptionalHeapPtr`.
 - `nstd_thread_sleep` now takes `NSTDDuration`.
 ### `nstd.time`
 - `nstd.time` functions now take `NSTDTime` & `NSTDDuration` by value.